+++ CHANGELOG.md
@@ -8,6 +8,11 @@
   on the Erlang target.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
+- Type mismatch errors now emphasize the exact location of mismatches in complex
+  types by highlighting the mismatched type parameters. This makes it easier to
+  spot errors in types with multiple parameters or deep nesting.
+  ([Adi Salimgereyev](https://github.com/abs0luty))
+
 ### Build tool
 
 - The help text displayed by `gleam dev --help`, `gleam test --help`, and